This commit was generated by cvs2svn to compensate for changes in r11851, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@11852 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/printerdrv/drv-mps803.c
@@ -27,6 +27,7 @@
 
 #include "vice.h"
 
+#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
@@ -75,17 +76,17 @@ static log_t drv803_log = LOG_ERR;
 /* ------------------------------------------------------------------------- */
 /* MPS803 printer engine. */
 
-static void set_mode(mps_t *mps, int m)
+static void set_mode(mps_t *mps, unsigned int m)
 {
     mps->mode |= m;
 }
 
-static void del_mode(mps_t *mps, int m)
+static void del_mode(mps_t *mps, unsigned int m)
 {
     mps->mode &= ~m;
 }
 
-static int is_mode(mps_t *mps, int m)
+static int is_mode(mps_t *mps, unsigned int m)
 {
     return mps->mode & m;
 }
@@ -131,9 +132,9 @@ static void write_line(mps_t *mps, unsigned int prnr)
 
     for (y = 0; y < 7; y++) {
         for (x = 0; x < 480; x++)
-            output_select_putc(prnr, mps->line[x][y] ? OUTPUT_PIXEL_BLACK
-                               : OUTPUT_PIXEL_WHITE);
-        output_select_putc(prnr, OUTPUT_NEWLINE);
+            output_select_putc(prnr, (BYTE)(mps->line[x][y]
+                               ? OUTPUT_PIXEL_BLACK : OUTPUT_PIXEL_WHITE));
+        output_select_putc(prnr, (BYTE)(OUTPUT_NEWLINE));
     }
 
     if (!is_mode(mps, MPS_BITMODE)) {
